% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.dictionary.R
\docType{data}
\name{animals.dictionary}
\alias{animals.dictionary}
\title{Animals Dictionary}
\format{
animals.dictionary (vector, length = 1211)
}
\usage{
data(animals.dictionary)
}
\description{
A database of possible animals responses (\emph{n} = 1211)
}
\details{
To add additional animals to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("animals.dictionary")
}
\keyword{datasets}
