% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.misnomer.R
\docType{data}
\name{animals.misnomer}
\alias{animals.misnomer}
\title{Animals Misnomer}
\format{animals.misnomer (list, length = 208)}
\usage{
data(animals.misnomer)
}
\description{
A database of possible animals misnomers, monikers, and common spelling errors
}
\details{
To add additional animals misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("animals.misnomer")
}
\keyword{datasets}
