% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.moniker.R
\docType{data}
\name{animals.moniker}
\alias{animals.moniker}
\title{Animals Moniker}
\format{
animals.moniker (list, length = 236)
}
\usage{
data(animals.moniker)
}
\description{
A database of possible animals monikers and common spelling errors
}
\details{
To add additional animals monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("animals.moniker")
}
\keyword{datasets}
