% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coca.dictionary.R
\docType{data}
\name{coca.dictionary}
\alias{coca.dictionary}
\title{Corpus of Contemporary American English Dictionary}
\format{
coca.dictionary (vector, length = 80381)
}
\usage{
data(coca.dictionary)
}
\description{
A general dictionary of over 80,000 words from the Corpus of Contemporary American English derived from
\href{https://www.wordfrequency.info/samples.asp}{https://www.wordfrequency.info/samples.asp}.
}
\details{
To add additional words to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("coca.dictionary")
}
\keyword{datasets}
