% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coca.moniker.R
\docType{data}
\name{coca.moniker}
\alias{coca.moniker}
\title{Corpus of Contemporary American English Moniker}
\format{
coca.moniker (list, length = 20267)
}
\usage{
data(coca.moniker)
}
\description{
A database of word forms for the Corpus of Contemporary American English dictionary
}
\details{
To add additional COCA monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("coca.moniker")
}
\keyword{datasets}
