% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocaspell.dictionary.R
\docType{data}
\name{cocaspell.dictionary}
\alias{cocaspell.dictionary}
\title{Corpus of Contemporary American English and Hunspell Combined Dictionary}
\format{
cocaspell.dictionary (vector, length = 109169)
}
\usage{
data(cocaspell.dictionary)
}
\description{
A general dictionary of over 109,000 words from the Corpus of Contemporary American English dictionary
(\code{\link[SemNetDictionaries]{coca.dictionary}}) and Hunspell dictionary (\code{\link[SemNetDictionaries]{hunspell.dictionary}}).
}
\details{
To add additional words to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("cocaspell.dictionary")
}
\keyword{datasets}
