% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\docType{data}
\name{dictionaries}
\alias{dictionaries}
\title{Dictionaries}
\format{dictionaries (vector)}
\usage{
data(dictionaries)
}
\description{
A list of dictionaries in \code{\link{SemNetDictionaries}}
}
\details{
To add additional dictionaries, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("dictionaries")
}
\keyword{datasets}
