% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.dictionaries.R
\name{extra.dictionaries}
\alias{extra.dictionaries}
\title{User-defined Dictionary List}
\usage{
extra.dictionaries()
}
\value{
Returns a vector of dictionaries that are not hard coded into
the \code{\link{SemNetDictionaries}} package
}
\description{
A wrapper function designed to identify the list of
user-defined dictionaries
(e.g., \code{\link[SemNetDictionaries]{append.dictionary}}).
}
\examples{

extra.dictionaries()

}
\seealso{
\code{\link{find.dictionaries}} to find where dictionaries are stored,
\code{\link{dictionaries}} to identify available dictionaries,
\code{\link{append.dictionary}} to create a new dictionary
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
