% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.dictionaries.R
\name{find.dictionaries}
\alias{find.dictionaries}
\title{Finds Names and Locations of Appendix Dictionaries}
\usage{
find.dictionaries()
}
\value{
\item{names}{Returns the names of the appendix dictionary file(s)}

\item{files}{Returns the dictionary file(s) that are stored in each given path. If there is no output
(e.g., \code{character(0)}), then no appendix dictionary file exists
(one can be created using the \code{\link[SemNetDictionaries]{append.dictionary}} function)}
}
\description{
A wrapper function to identify the save location
of appendix dictionaries from \code{\link[SemNetDictionaries]{append.dictionary}}
}
\details{
This function identifies where the
\code{\link[SemNetDictionaries]{append.dictionary}} function
has stored appendix dictionaries. The default for the
\code{\link[SemNetDictionaries]{append.dictionary}} function
is to store dictionaries in your temporary \code{R} data folder.
If you do not know where this is, a path is provided to you
in the output of this function (see \code{files} below).
}
\examples{
\dontrun{

#No appendix dictionaries found
find.dictionaries()

#Make a dictionary
append.dictionary(c("words","are","fun"))

#Dictionary can now be found
find.dictionaries()
}

}
\seealso{
\code{\link{append.dictionary}} to create a new dictionary,
\code{\link{dictionaries}} to identify dictionaries in
\code{\link{SemNetDictionaries}}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
