% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruits.dictionary.R
\docType{data}
\name{fruits.dictionary}
\alias{fruits.dictionary}
\title{Fruits Dictionary}
\format{fruits.dictionary (vector, length = 488)}
\usage{
data(fruits.dictionary)
}
\description{
A database of possible fruits responses (\emph{n} = 488)
}
\details{
To add additional fruits to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("fruits.dictionary")
}
\keyword{datasets}
