% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruits.misnomer.R
\docType{data}
\name{fruits.misnomer}
\alias{fruits.misnomer}
\title{Fruits Misnomer}
\format{fruits.misnomer (list, length = 39)}
\usage{
data(fruits.misnomer)
}
\description{
A database of possible fruits misnomers, monikers, and common spelling errors
}
\details{
To add additional fruits misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("fruits.misnomer")
}
\keyword{datasets}
