% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fruits.moniker.R
\docType{data}
\name{fruits.moniker}
\alias{fruits.moniker}
\title{Fruits Moniker}
\format{fruits.moniker (list, length = 39)}
\usage{
data(fruits.moniker)
}
\description{
A database of possible fruits monikers and common spelling errors
}
\details{
To add additional fruits monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("fruits.moniker")
}
\keyword{datasets}
