% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.dictionary.R
\docType{data}
\name{general.dictionary}
\alias{general.dictionary}
\title{General Dictionary}
\format{general.dictionary (vector, length = 455298)}
\usage{
data(general.dictionary)
}
\description{
A general dictionary of over 400,000 words (\emph{n} = 455298) derived from
\href{https://github.com/dwyl/english-words}{https://github.com/dwyl/english-words}.
All punctuations have been removed.
}
\details{
To add additional animals to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("general.dictionary")
}
\keyword{datasets}
