% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good.dictionary.R
\docType{data}
\name{good.dictionary}
\alias{good.dictionary}
\title{'Good' Synonyms Dictionary}
\format{
good.dictionary (vector, length = 284)
}
\usage{
data(good.dictionary)
}
\description{
A database of possible good synonym responses (\emph{n} = 284) 
To add additional good synonyms to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("good.dictionary")
}
\keyword{datasets}
