% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good.misnomer.R
\docType{data}
\name{good.misnomer}
\alias{good.misnomer}
\title{'Good' Misnomer}
\format{good.misnomer (list, length = 4)}
\usage{
data(good.misnomer)
}
\description{
A database of possible good misnomers, monikers, and common spelling errors
}
\details{
To add additional good misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("good.misnomer")
}
\keyword{datasets}
