% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/good.moniker.R
\docType{data}
\name{good.moniker}
\alias{good.moniker}
\title{'Good' Moniker}
\format{good.moniker (list, length = 4)}
\usage{
data(good.moniker)
}
\description{
A database of possible good monikers and common spelling errors
}
\details{
To add additional good monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("good.moniker")
}
\keyword{datasets}
