% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.dictionary.R
\docType{data}
\name{hot.dictionary}
\alias{hot.dictionary}
\title{'Hot' Synonyms Dictionary}
\format{hot.dictionary (vector, length = 281)}
\usage{
data(hot.dictionary)
}
\description{
A database of possible hot synonym responses (\emph{n} = 281) 
To add additional hot synonyms to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("hot.dictionary")
}
\keyword{datasets}
