% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.misnomer.R
\docType{data}
\name{hot.misnomer}
\alias{hot.misnomer}
\title{Hot Misnomer}
\format{hot.misnomer (list, length = 15)}
\usage{
data(hot.misnomer)
}
\description{
A database of possible hot misnomers, monikers, and common spelling errors
}
\details{
To add additional hot misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("hot.misnomer")
}
\keyword{datasets}
