% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hot.moinker.R
\docType{data}
\name{hot.moniker}
\alias{hot.moniker}
\title{Hot Moniker}
\format{hot.moniker (list, length = 15)}
\usage{
data(hot.moniker)
}
\description{
A database of possible hot monikers and common spelling errors
}
\details{
To add additional hot monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("hot.moniker")
}
\keyword{datasets}
