% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunspell.dictionary.R
\docType{data}
\name{hunspell.dictionary}
\alias{hunspell.dictionary}
\title{hunspell Dictionary}
\format{
hunspell.dictionary (vector, length = 38530)
}
\usage{
data(hunspell.dictionary)
}
\description{
A general dictionary of over 38,000 words from the hunspell dictionary derived from
\href{http://wordlist.aspell.net/dicts/}{http://wordlist.aspell.net/dicts/}.
}
\details{
To add additional words to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("hunspell.dictionary")
}
\keyword{datasets}
