% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.dictionary.R
\docType{data}
\name{jobs.dictionary}
\alias{jobs.dictionary}
\title{Jobs Dictionary}
\format{jobs.dictionary (vector, length = 1471)}
\usage{
data(jobs.dictionary)
}
\description{
A database of possible jobs and related words (\emph{n} = 1471)
}
\details{
To add additional jobs to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("jobs.dictionary")
}
\keyword{datasets}
