% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.misnomer.R
\docType{data}
\name{jobs.misnomer}
\alias{jobs.misnomer}
\title{Jobs Misnomer}
\format{jobs.misnomer (list, length = 117)}
\usage{
data(jobs.misnomer)
}
\description{
A database of possible jobs misnomers, monikers, and common spelling errors
}
\details{
To add additional jobs misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("jobs.misnomer")
}
\keyword{datasets}
