% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.moniker.R
\docType{data}
\name{jobs.moniker}
\alias{jobs.moniker}
\title{Jobs Moniker}
\format{jobs.moniker (list, length = 117)}
\usage{
data(jobs.moniker)
}
\description{
A database of possible jobs monikers and common spelling errors
}
\details{
To add additional jobs monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("jobs.moniker")
}
\keyword{datasets}
