% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.misnomers.R
\name{load.misnomers}
\alias{load.misnomers}
\title{Load Misnomers}
\usage{
load.misnomers(misnomer)
}
\arguments{
\item{misnomer}{Character vector.
Misnomers to load (must be a dictionary in
\code{\link[SemNetDictionaries]{dictionaries}})}
}
\value{
Returns a vector of unique words that have been combined
and alphabetized from the specified misnomers
}
\description{
A wrapper function to load misnomers into
the 'SemNetCleaner' package. Searches for misnomers in \code{R}'s
 \code{\link{SemNetDictionaries}} package. Outputs a unique word list
that is combined from all dictionaries entered in the \code{misnomer} argument
}
\examples{
#find dictionaries to load
dictionaries()

#load "animals" misnomers
load.misnomers("animals")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
