% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_common.dictionary.R
\docType{data}
\name{most_common.dictionary}
\alias{most_common.dictionary}
\title{Most Common Dictionary}
\format{
most_common.dictionary (vector, length = 9329)
}
\usage{
data(most_common.dictionary)
}
\description{
A general dictionary of 10,000 of the most common U.S. English words derived from
\href{https://github.com/first20hours/google-10000-english}{https://github.com/first20hours/google-10000-english}.
}
\details{
To add additional words to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("most_common.dictionary")
}
\keyword{datasets}
