% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemNetDictionaries.R
\docType{package}
\name{SemNetDictionaries-package}
\alias{SemNetDictionaries}
\alias{SemNetDictionaries-package}
\title{SemNetDictionaries--package}
\description{
Implements dictionaries that can be used in the \code{\link[SemNetCleaner]{SemNetCleaner-package}}.
Also includes several functions aimed at facilitating the text cleaning analysis in the \code{\link[SemNetCleaner]{SemNetCleaner-package}}.
This package is designed to integrate and update word lists and dictionaries based on each
user's individual needs by allowing users to store and save their own dictionaries.
Dictionaries can be added to the \code{\link{SemNetDictionaries}} package by submitting user-defined
dictionaries to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexChristensen/SemNetDictionaries}
  \item Report bugs at \url{https://github.com/AlexChristensen/SemNetDictionaries/issues}
}

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
