% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWoRdle.R
\name{ShinyWoRdle}
\alias{ShinyWoRdle}
\title{Shiny App to Play \href{WoRdle}{https://www.powerlanguage.co.uk/wordle/}}
\usage{
ShinyWoRdle()
}
\description{
An interactive Shiny application for playing \href{WoRdle}{https://www.powerlanguage.co.uk/wordle/}
}
\examples{

if(interactive())
{ShinyWoRdle()}


}
