% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.dictionary.R
\docType{data}
\name{stop_words.dictionary}
\alias{stop_words.dictionary}
\title{Stop Words Dictionary}
\format{
stop_words.dictionary (vector, length = 56)
}
\usage{
data(stop_words.dictionary)
}
\description{
A selection of stop words that can be removed from semantic responses (\emph{n} = 56)
}
\details{
To add additional animals to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("stop_words.dictionary")
}
\keyword{datasets}
