% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetables.dictionary.R
\docType{data}
\name{vegetables.dictionary}
\alias{vegetables.dictionary}
\title{Vegetables Dictionary}
\format{vegetables.dictionary (vector, length = 284)}
\usage{
data(vegetables.dictionary)
}
\description{
A database of possible vegetables responses (\emph{n} = 284)
}
\details{
To add additional vegetables to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("vegetables.dictionary")
}
\keyword{datasets}
