% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetables.misnomer.R
\docType{data}
\name{vegetables.misnomer}
\alias{vegetables.misnomer}
\title{Vegetables Misnomer}
\format{vegetables.misnomer (list, length = 35)}
\usage{
data(vegetables.misnomer)
}
\description{
A database of possible vegetables misnomers, monikers, and common spelling errors
}
\details{
To add additional vegetables misnomers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("vegetables.misnomer")
}
\keyword{datasets}
