% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetables.moniker.R
\docType{data}
\name{vegetables.moniker}
\alias{vegetables.moniker}
\title{Vegetables Moniker}
\format{vegetables.moniker (list, length = 35)}
\usage{
data(vegetables.moniker)
}
\description{
A database of possible vegetables monikers and common spelling errors
}
\details{
To add additional vegetables monikers to the database, please submit a pull
request or issue to \href{https://github.com/AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{
data("vegetables.moniker")
}
\keyword{datasets}
