% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.dictionary.R
\name{append.dictionary}
\alias{append.dictionary}
\title{Appendix Dictionary}
\usage{
append.dictionary(..., dict.name)
}
\arguments{
\item{...}{Character vector.
A vector of words to create or add to dictionary}

\item{dict.name}{Characters.
Name of dictionary to create or add words to.
Defaults to \code{"appendix.dictionary"}.
Input a name to create or add to an existing dictionary}
}
\description{
A function designed to post-hoc update the dictionaries in the
\code{\link{SemNetDictionaries}}. This function is used to add
words that are not included in \code{\link{SemNetDictionaries}}'
dictionaries. This allows for new semantic categories or words to be included
for future use. THIS IS YOUR OWN PERSONAL DICTIONARY. Dictionaries created using this function
will ONLY be available on your current computer. An open-source community-derived
dictionaries can be uploaded to and downloaded from
\href{AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\details{
Dictionaries are stored in your \code{R} packages library.
To find where your dictionaries are stored, use the
\code{\link[SemNetDictionaries]{find.dictionaries}} function. A new
folder is created, so that your appendix dictionary files can be stored
without deletion when updating the \code{\link{SemNetDictionaries}}
package. These dictionaries are only stored on your private computer
and must either be publicly shared or transferred to other computers in
order to use them elsewhere. Please see the documentation for
\code{\link{SemNetDictionaries}} for more details.

Appendix dictionaries are useful for storing spelling
definitions that are not available in the \code{\link{SemNetDictionaries}}
package. This function enables the storage of personalized dictionaries,
which can be used in combination with other dictionaries to facilitate
the cleaning of text data. Although these dictionaries are personal,
they can also be shared for other users. Please submit a pull request
with your dictionary to be added to the package at: \href{AlexChristensen/SemNetDictionaries}{https://github.com/AlexChristensen/SemNetDictionaries}
}
\examples{

append.dictionary(c("words","are","fun"))

}
\seealso{
\code{\link{find.dictionaries}} to find where dictionaries are stored,
\code{\link{dictionaries}} to identify available dictionaries,
\code{\link{extra.dictionaries}} to see appendix dictionaries not in
\code{\link{SemNetDictionaries}}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
