% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brit2us.R
\docType{data}
\name{brit2us}
\alias{brit2us}
\title{British-US English Conversions}
\format{
brit2us (list, length = 780)
}
\usage{
data(brit2us)
}
\description{
A database to convert between British and US spellings (\emph{n} = 780)
}
\examples{
data("brit2us")
}
\keyword{datasets}
