% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dictionaries}
\alias{dictionaries}
\title{List Names of Dictionaries in 'SemNetDictionaries'}
\usage{
dictionaries(quiet)
}
\arguments{
\item{quiet}{Boolean.
Determines whether the return should be quiet (does not print dictionaries).
Defaults to \code{FALSE}}
}
\value{
Returns the names of dictionaries in \code{\link{SemNetDictionaries}}
}
\description{
A wrapper function to identify all dictionaries included in 
\code{\link{SemNetDictionaries}}
}
\examples{
# List names of dictionaries in 'SemNetDictionaries'
dictionaries()

}
\seealso{
\code{\link{find.dictionaries}} to find where dictionaries are stored,
\code{\link{append.dictionary}} to create a new dictionary
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
