% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.dictionaries.R
\name{find.dictionaries}
\alias{find.dictionaries}
\title{Finds Locations of Dictionaries}
\usage{
find.dictionaries()
}
\value{
Returns a list containing:

\item{location}{The location where \code{R} saves the
dictionaries}

\item{files}{The dictionary files that are stored in the location. If there is no output
(e.g., \code{character(0)}), then no appendix dictionary file exists
(one can be created using the \code{\link[SemNetDictionaries]{append.dictionary}} function)}
}
\description{
A wrapper function to identify the save location
of appendix dictionaries from \code{\link[SemNetDictionaries]{append.dictionary}}
}
\details{
This function identifies where the
\code{\link[SemNetDictionaries]{append.dictionary}} function
has stored appendix dictionaries. The default for the
\code{\link[SemNetDictionaries]{append.dictionary}} function
is to store dictionaries in your \code{R} packages folder.
This folder is the default for where \code{R} stores your downloaded and installed packages.
}
\examples{

find.dictionaries()

}
\seealso{
\code{\link{append.dictionary}} to create a new dictionary,
\code{\link{dictionaries}} to identify available dictionaries,
\code{\link{extra.dictionaries}} to see appendix dictionaries not in
\code{\link{SemNetDictionaries}}
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
