% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.dictionary.R
\docType{data}
\name{general.dictionary}
\alias{general.dictionary}
\title{General Dictionary}
\format{general.dictionary (vector, length = 370103)}
\usage{
data(general.dictionary)
}
\description{
A general dictionary of over 370,000 words (\emph{n} = 370,103) derived from
\href{https://github.com/dwyl/english-words}{https://github.com/dwyl/english-words}.
All punctuation have been removed.
}
\details{
To add additional words to the dictionary, please make an
appendix dictionary (\code{\link[SemNetDictionaries]{append.dictionary}})
}
\examples{
data("general.dictionary")
}
\keyword{datasets}
