% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.dictionaries.R
\name{load.dictionaries}
\alias{load.dictionaries}
\title{Load Dictionaries}
\usage{
load.dictionaries(dictionary)
}
\arguments{
\item{dictionary}{Character vector.
Dictionaries to load (see examples)}
}
\value{
Returns a vector of words in specified dictionaries
}
\description{
A wrapper function for to load text dictionaries in
\code{\link{SemNetCleaner}}
}
\examples{

load.dictionaries("animals")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
