% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWoRdle.R
\name{ShinyWoRdle}
\alias{ShinyWoRdle}
\title{Shiny App to Play WoRdle}
\usage{
ShinyWoRdle()
}
\description{
An interactive Shiny application for playing WoRdle
}
\examples{

if(interactive())
{ShinyWoRdle()}


}
